+++ docs/changelog.md
@@ -1,14 +1,22 @@
 # Changelog
 
-## [v0.2.2](https://github.com/0x2142/frigate-notify/releases/tag/v0.2.2)
+## [v0.2.3](https://github.com/0x2142/frigate-notify/releases/tag/v0.2.3) - Upcoming Release
+
+- Update config management
+- Improve config file validation
+
+## [v0.2.2](https://github.com/0x2142/frigate-notify/releases/tag/v0.2.2) - Nov 29 2023
+
  - Fix SMTP issue where code was ignoring `tls: false` config flag
  - Fix issue where snapshots would only be sent to the first alerting method, if multiple were enabled 
  - Update Debian image used for Docker image 
 
-## [v0.2.1](https://github.com/0x2142/frigate-notify/releases/tag/v0.2.1)
+## [v0.2.1](https://github.com/0x2142/frigate-notify/releases/tag/v0.2.1) - Sep 20 2023
+
  - Added Docker release build for ARM
 
-## [v0.2.0](https://github.com/0x2142/frigate-notify/releases/tag/v0.2.0)
+## [v0.2.0](https://github.com/0x2142/frigate-notify/releases/tag/v0.2.0) - Sep 19 2023
+
  - Add support for excluding cameras from alerts
  - Added ability to set custom alert titles
  - Added ability to set custom MQTT client ID
@@ -18,8 +26,10 @@
  - Add GitHub actions for automated release builds
  - Added version number to startup logging
 
-## [v0.1.1](https://github.com/0x2142/frigate-notify/releases/tag/v0.1.1)
+## [v0.1.1](https://github.com/0x2142/frigate-notify/releases/tag/v0.1.1) - Apr 25 2023
+
  - Add support for MQTT authentication
 
-## [v0.1.0](https://github.com/0x2142/frigate-notify/releases/tag/v0.1.0)
+## [v0.1.0](https://github.com/0x2142/frigate-notify/releases/tag/v0.1.0) - Mar 17 2023
+
  - Initial release! 🎉